requirements.txt
numpy==1.21.6
pandas==1.3.5
openpyxl==3.0.9
opencv-python==4.5.5.64
Pillow==8.4.0

